+++ CHANGELOG.md
@@ -7,6 +7,8 @@ format. This project adheres to [Semantic Versioning](https://semver.org/spec/v2
 
 ### Fixed
 
+- Accept `URI` structs for the URL and the default value in
+  `SafeRedirect.redirect/4`.
 - Raise `ArgumentError` instead of `CaseClauseError` if the resolved URL cannot
   be redirected to.